Разбивает utils.js на несколько модулей
