+++ src/utils/render.js
@@ -1,0 +1,81 @@
+import Abstract from "../view/abstract.js";
+
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
+};
+
+const render = (container, child, place) => {
+  if (container instanceof Abstract) {
+    container = container.element;
+  }
+
+  if (child instanceof Abstract) {
+    child = child.element;
+  }
+
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(child);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(child);
+      break;
+    case RenderPosition.AFTEREND:
+      container.after(child);
+      break;
+  }
+};
+
+const renderTemplate = (container, template, place) => {
+  if (container instanceof Abstract) {
+    container = container.element;
+  }
+
+  container.insertAdjacentHTML(place, template);
+};
+
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+const replace = (newChild, oldChild) => {
+  if (oldChild instanceof Abstract) {
+    oldChild = oldChild.element;
+  }
+
+  if (newChild instanceof Abstract) {
+    newChild = newChild.element;
+  }
+
+  const parent = oldChild.parentElement;
+
+  if (parent === null || oldChild === null || newChild === null) {
+    throw new Error(`Can't replace unexisting elements`);
+  }
+
+  parent.replaceChild(newChild, oldChild);
+};
+
+const remove = (component) => {
+  if (!(component instanceof Abstract)) {
+    throw new Error(`Can remove only components`);
+  }
+
+  component.element.remove();
+  component.element = null;
+};
+
+export {
+  RenderPosition,
+  render,
+  renderTemplate,
+  createElement,
+  replace,
+  remove
+};
